% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_var_function.R
\name{extract_var}
\alias{extract_var}
\title{A Function to extract variables}
\usage{
extract_var(data)
}
\description{
This function allows you to extract the variables from the bikes data.
Make new columns for the day, month and hour from the datetime column.
And also set day, month and hour as factors.
}
\examples{
extract_var()
}
