% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_all_function.R
\name{keep_all}
\alias{keep_all}
\title{A Function to extract variablesa and count}
\usage{
keep_all(data)
}
\description{
This function allows you to extract the variables from the bikes dataset and include the column count.
Make new columns for the day, month and hour from the datetime column.
And also set day, month and hour as factors.
}
\examples{
keep_all()
}
